improved styling and mobile view
